requirements.txt
aioredis==2.0.1
amqp==5.3.1
annoy==1.17.3
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
black==25.1.0
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.5
cryptography==45.0.6
daphne==4.2.0
Django==5.1.8
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-fernet-fields==0.6
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isort==6.0.1
joblib==1.5.2
kombu==5.5.4
mongoengine==0.29.1
mypy_extensions==1.1.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pymongo==4.14.1
pyOpenSSL==25.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==5.2.1
scikit-learn==1.7.2
scipy==1.16.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
